# In-order iteration for rowSums2()